pyproject.toml
[tool.poetry]
name = "investors-diary"
version = "0.1.0"
description = "An LLM Agent based personal diary that manages the purchase and selling of the stocks using natural language."
authors = ["Kirushikesh DB"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
langgraph = "^0.2.50"
langchain-openai = "^0.2.8"
jupyter = "^1.1.1"
python-dotenv = "^1.0.1"
langchain = "^0.3.7"
langchain-community = "^0.3.7"
faiss-cpu = "^1.9.0"
tavily-python = "^0.5.0"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
langserve = {extras = ["all"], version = "^0.3.0"}
streamlit = "^1.40.1"


[tool.poetry.group.lint.dependencies]
ruff = ">0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
